# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Urva Gandhi Portfolio Developer API & MCP Server",
    "description": "Public REST APIs, Model Context Protocol (MCP) server, and developer endpoints providing verified coding statistics, developer profile metadata, and interactive contact handling for Urva Gandhi.",
    "version": "1.0.0",
    "contact": {
      "name": "Urva Gandhi",
      "email": "23bce078@nirmauni.ac.in",
      "url": "https://urvagandhi.tech"
    },
    "x-versioning-policy": "URL path versioning (/v1/) and X-API-Version header parameter. Backward compatibility guaranteed for version 1.",
    "x-deprecation-policy": "Deprecation notices served via Sunset HTTP header 6 months prior to end-of-life."
  },
  "servers": [
    {
      "url": "https://urvagandhi.tech",
      "description": "Primary Production Domain"
    },
    {
      "url": "https://urvagandhi-portfolio.vercel.app",
      "description": "Lifetime Vercel Mirror Domain"
    },
    {
      "url": "https://urvagandhi.tech/v1",
      "description": "Active Domain Versioned Base URL (/v1)"
    }
  ],
  "paths": {
...
    "/api/gfg": {
      "get": {
        "operationId": "getGfgStats",
        "summary": "Get GeeksforGeeks Statistics",
        "description": "Fetches GeeksforGeeks overall coding score, streak data, and institute ranking.",
        "security": [],
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version to target. Omit to use the latest stable version (1).",
            "schema": {
              "type": "string",
              "enum": ["1"],
              "default": "1"
            }
          },
          {
            "name": "username",
            "in": "query",
            "required": false,
            "description": "GeeksforGeeks handle (defaults to urvagandhi)",
            "schema": {
              "type": "string",
              "default": "urvagandhi"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful GeeksforGeeks response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GfgProfile"
                }
              }
            }
          }
        }
      }
    },
    "/api/hackerrank": {
      "get": {
        "operationId": "getHackerRankStats",
        "summary": "Get HackerRank Statistics",
        "description": "Fetches HackerRank badges, scores, and problem solving achievements.",
        "security": [],
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version to target. Omit to use the latest stable version (1).",
            "schema": {
              "type": "string",
              "enum": ["1"],
              "default": "1"
            }
          },
          {
            "name": "username",
            "in": "query",
            "required": false,
            "description": "HackerRank handle (defaults to urvagandhi24)",
            "schema": {
              "type": "string",
              "default": "urvagandhi24"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful HackerRank response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HackerRankProfile"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/gfg": {
      "get": {
        "operationId": "getGfgStats",
        "summary": "Get GeeksforGeeks Statistics",
        "description": "Fetches GeeksforGeeks overall coding score, streak data, and institute ranking.",
        "security": [],
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version to target. Omit to use the latest stable version (1).",
            "schema": {
              "type": "string",
              "enum": ["1"],
              "default": "1"
            }
          },
          {
            "name": "username",
            "in": "query",
            "required": false,
            "description": "GeeksforGeeks handle (defaults to urvagandhi)",
            "schema": {
              "type": "string",
              "default": "urvagandhi"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful GeeksforGeeks response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GfgProfile"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gfg": {
      "get": {
        "operationId": "getGfgStats",
        "summary": "Get GeeksforGeeks Statistics",
        "description": "Fetches GeeksforGeeks overall coding score, streak data, and institute ranking.",
        "security": [],
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version to target. Omit to use the latest stable version (1).",
            "schema": {
              "type": "string",
              "enum": ["1"],
              "default": "1"
            }
          },
          {
            "name": "username",
            "in": "query",
            "required": false,
            "description": "GeeksforGeeks handle (defaults to urvagandhi)",
            "schema": {
              "type": "string",
              "default": "urvagandhi"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful GeeksforGeeks response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GfgProfile"
                }
              }
            }
          }
        }
      }
    },
    "/v1/hackerrank": {
      "get": {
        "operationId": "getHackerRankStats",
        "summary": "Get HackerRank Statistics",
        "description": "Fetches HackerRank badges, scores, and problem solving achievements.",
        "security": [],
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version to target. Omit to use the latest stable version (1).",
            "schema": {
              "type": "string",
              "enum": ["1"],
              "default": "1"
            }
          },
          {
            "name": "username",
            "in": "query",
            "required": false,
            "description": "HackerRank handle (defaults to urvagandhi24)",
            "schema": {
              "type": "string",
              "default": "urvagandhi24"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful HackerRank response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HackerRankProfile"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hackerrank": {
      "get": {
        "operationId": "getHackerRankStats",
        "summary": "Get HackerRank Statistics",
        "description": "Fetches HackerRank badges, scores, and problem solving achievements.",
        "security": [],
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version to target. Omit to use the latest stable version (1).",
            "schema": {
              "type": "string",
              "enum": ["1"],
              "default": "1"
            }
          },
          {
            "name": "username",
            "in": "query",
            "required": false,
            "description": "HackerRank handle (defaults to urvagandhi24)",
            "schema": {
              "type": "string",
              "default": "urvagandhi24"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful HackerRank response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HackerRankProfile"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://urvagandhi.tech/.well-known/oauth/token",
            "scopes": {
              "profile:read": "Read public developer profile and coding statistics.",
              "profile:write": "Submit a contact inquiry on behalf of an authenticated client."
            }
          },
          "authorizationCode": {
            "authorizationUrl": "https://urvagandhi.tech/.well-known/oauth/authorize",
            "tokenUrl": "https://urvagandhi.tech/.well-known/oauth/token",
            "scopes": {
              "profile:read": "Read public developer profile and coding statistics.",
              "profile:write": "Submit a contact inquiry on behalf of an authenticated client."
            }
          }
        }
      }
    },
    "schemas": {
...
      "GfgProfile": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "urvagandhi"
          },
          "score": {
            "type": "integer",
            "example": 450
          }
        }
      },
      "HackerRankProfile": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "urvagandhi24"
          }
        }
      },
...
    }
  }
}